+++ src/components/Icon/DeleteIcon.jsx
@@ -1,0 +1,30 @@
+import React from "react";
+
+const DeleteIcon = () => {
+  return (
+    <>
+      <svg
+        width="40"
+        height="40"
+        viewBox="0 0 40 40"
+        fill="none"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M26.7937 17.1134L25.8817 28.5934C25.8289 29.2061 25.3253 29.6675 24.7126 29.6675H15.4441C14.8279 29.6675 14.3278 29.2062 14.275 28.597L13.363 17.1135C13.3314 16.7261 12.9968 16.4374 12.6058 16.4656C12.2185 16.4972 11.9298 16.8354 11.958 17.2227L12.87 28.7097C12.9826 30.0585 14.0884 31.0762 15.4441 31.0762H24.7161C26.0683 31.0762 27.1742 30.0585 27.2903 28.7062L28.2023 17.2227C28.2339 16.8354 27.9452 16.4973 27.5544 16.4656C27.1635 16.4374 26.8254 16.7261 26.7938 17.1134L26.7937 17.1134Z"
+          fill="#222222"
+        />
+        <path
+          d="M16.4372 13.092C16.8245 13.092 17.1415 12.775 17.1415 12.3877V11.0002C17.1415 10.6727 17.4056 10.4086 17.7331 10.4086H21.8849C22.2124 10.4086 22.4765 10.6727 22.4765 11.0002V12.3877C22.4765 12.775 22.7935 13.092 23.1808 13.092C23.5681 13.092 23.8851 12.775 23.8851 12.3877V11.0002C23.8851 9.898 22.9871 9 21.8849 9H17.7331C16.6309 9 15.7329 9.89803 15.7329 11.0002V12.3877C15.7329 12.7786 16.0463 13.092 16.4372 13.092Z"
+          fill="#222222"
+        />
+        <path
+          d="M10.7043 15.564H28.9842C29.3715 15.564 29.6885 15.247 29.6885 14.8597C29.6885 14.4723 29.3715 14.1554 28.9842 14.1554H10.7043C10.317 14.1554 10 14.4723 10 14.8597C10 15.247 10.317 15.564 10.7043 15.564Z"
+          fill="#222222"
+        />
+      </svg>
+    </>
+  );
+};
+
+export default DeleteIcon;
